manager: Make FsManager and SystemdManager serializable
The kata-agent (de)serializes managers in child initialization. The `oci`
feature relies on `serde` crate to provide the ability to serialize.

Zbus's value uses variable reference, which is hard to serialize. `Value`
is introduced to take the ownership of the value, so that we can remove
lifetime annotations from `SystemdClient`, also from `SystemdManager`.

The types supported by the `Value` only include the types that are in use.
Later on, more types can be added to `Value` if needed.

Signed-off-by: Xuewei Niu <niuxuewei.nxw@antgroup.com>